djangoframework/requirements.txt
amqp==5.3.1
asgiref==3.11.0
attrs==25.4.0
billiard==4.2.4
celery==5.6.0
certifi==2025.11.12
channels==4.3.2
channels_redis==4.3.0
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
Django==6.0
django-filter==25.2
django-redis==6.0.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
exceptiongroup==1.3.1
Faker==37.11.0
h11==0.16.0
idna==3.11
kombu==5.6.1
msgpack==1.1.2
outcome==1.3.0.post0
packaging==25.0
prompt_toolkit==3.0.52
psycopg==3.2.9
PyJWT==2.10.1
pymemcache==4.0.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
redis==7.1.0
selenium==4.38.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.5.3
trio==0.32.0
trio-websocket==0.12.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
uvicorn==0.35.0
vine==5.1.0
wcwidth==0.2.14
websocket-client==1.9.0
wsproto==1.3.2
